Refactor `Cleanup`.

+++ Library/Homebrew/cleanup.rb
@@ -1,6 +1,71 @@
 require "utils/bottles"
 require "formula"
 
+module CleanupRefinement
+  refine Pathname do
+    def incomplete?
+      extname.end_with?(".incomplete")
+    end
+
+    def nested_cache?
+      directory? && ["glide_home", "java_cache", "npm_cache"].include?(basename.to_s)
+    end
+
+    def prune?(days)
+      return false unless days
+      return true if days.zero?
+
+      # TODO: Replace with ActiveSupport's `.days.ago`.
+      mtime < ((@time ||= Time.now) - days * 60 * 60 * 24)
+    end
+
+    def stale?(scrub = false)
+      return false unless file?
+
+      stale_formula?(scrub)
+    end
+
+    private
+
+    def stale_formula?(scrub)
+      return false unless HOMEBREW_CELLAR.directory?
+
+      version = if to_s.match?(Pathname::BOTTLE_EXTNAME_RX)
+        begin
+          Utils::Bottles.resolve_version(self)
+        rescue
+          self.version
+        end
+      else
+        self.version
+      end
+
+      return false unless version
+      return false unless (name = basename.to_s[/\A(.*?)\-\-?(?:#{Regexp.escape(version)})/, 1])
+
+      formula = begin
+        Formulary.from_rack(HOMEBREW_CELLAR/name)
+      rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
+        return false
+      end
+
+      if version.is_a?(PkgVersion)
+        return true if formula.pkg_version > version
+      elsif formula.version > version
+        return true
+      end
+
+      return true if scrub && !formula.installed?
+
+      return true if Utils::Bottles.file_outdated?(formula, self)
+
+      false
+    end
+  end
+end
+
+using CleanupRefinement
+
 module Homebrew
   module Cleanup
     @disk_cleanup_size = 0
@@ -43,25 +108,22 @@ module Homebrew
       unremovable_kegs << keg
     end
 
+    DEFAULT_LOG_DAYS = 14
+
     def cleanup_logs
       return unless HOMEBREW_LOGS.directory?
       HOMEBREW_LOGS.subdirs.each do |dir|
-        cleanup_path(dir) { dir.rmtree } if prune?(dir, days_default: 14)
+        cleanup_path(dir) { dir.rmtree } if dir.prune?(ARGV.value("prune")&.to_i || DEFAULT_LOG_DAYS)
       end
     end
 
     def cleanup_cache(cache = HOMEBREW_CACHE)
       return unless cache.directory?
       cache.children.each do |path|
-        if path.to_s.end_with? ".incomplete"
-          cleanup_path(path) { path.unlink }
-          next
-        end
-        if %w[glide_home java_cache npm_cache].include?(path.basename.to_s) && path.directory?
-          cleanup_path(path) { FileUtils.rm_rf path }
-          next
-        end
-        if prune?(path)
+        next cleanup_path(path) { path.unlink } if path.incomplete?
+        next cleanup_path(path) { FileUtils.rm_rf path } if path.nested_cache?
+
+        if path.prune?(ARGV.value("prune")&.to_i)
           if path.file?
             cleanup_path(path) { path.unlink }
           elsif path.directory? && path.to_s.include?("--")
@@ -70,42 +132,13 @@ module Homebrew
           next
         end
 
-        next unless path.file?
-        file = path
-
-        if file.to_s =~ Pathname::BOTTLE_EXTNAME_RX
-          version = begin
-                      Utils::Bottles.resolve_version(file)
-                    rescue
-                      file.version
-                    end
-        else
-          version = file.version
-        end
-        next unless version
-        next unless (name = file.basename.to_s[/\A(.*?)\-\-?(?:#{Regexp.escape(version)})/, 1])
-
-        next unless HOMEBREW_CELLAR.directory?
-
-        begin
-          f = Formulary.from_rack(HOMEBREW_CELLAR/name)
-        rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
-          next
-        end
-
-        file_is_stale = if version.is_a?(PkgVersion)
-          f.pkg_version > version
-        else
-          f.version > version
-        end
-
-        if file_is_stale || ARGV.switch?("s") && !f.installed? || Utils::Bottles.file_outdated?(f, file)
-          cleanup_path(file) { file.unlink }
-        end
+        next cleanup_path(path) { path.unlink } if path.stale?(ARGV.switch?("s"))
       end
     end
 
     def cleanup_path(path)
+      disk_usage = path.disk_usage
+
       if ARGV.dry_run?
         puts "Would remove: #{path} (#{path.abv})"
       else
@@ -113,7 +146,7 @@ module Homebrew
         yield
       end
 
-      update_disk_cleanup_size(path.disk_usage)
+      update_disk_cleanup_size(disk_usage)
     end
 
     def cleanup_lockfiles
@@ -142,27 +175,6 @@ module Homebrew
         end
       end
       workers.map(&:join)
-    end
-
-    def prune?(path, options = {})
-      @time ||= Time.now
-
-      path_modified_time = path.mtime
-      days_default = options[:days_default]
-
-      prune = ARGV.value "prune"
-
-      return true if prune == "all"
-
-      prune_time = if prune
-        @time - 60 * 60 * 24 * prune.to_i
-      elsif days_default
-        @time - 60 * 60 * 24 * days_default.to_i
-      end
-
-      return false unless prune_time
-
-      path_modified_time < prune_time
     end
   end
 end
